rename error folder to errors
