+++ frontEnd/src/components/errors/TeamNotFoundError.ts
