+++ frontEnd/src/components/errors/UnauthorizedError.ts
